create CountryDetailsInfor
